+++ src/opi/output/models/json/gbw/properties/molecule.py
@@ -7,6 +7,7 @@ from opi.output.models.json.gbw.properties.molecular_orbitals import (
     MolecularOrbitals,
 )
 from opi.output.models.json.gbw.properties.tddft import TdDft
+from opi.output.models.json.gbw.properties.two_electron_integrals import TwoElectronIntegrals
 
 
 class Molecule(GetItem):
@@ -33,12 +34,20 @@ class Molecule(GetItem):
         Overlap matrix
     h_matrix: list[list[StrictFloat]]
         Hcore matrix (1-el integrals)
+    t_matrix: list[list[StrictFloat]]
+        Kinetic energy matrix (1-el integrals)
+    v_matrix: list[list[StrictFloat]]
+        Nuclear attraction matrix (1-el integrals)
+    hmo: list[list[list[StrictFloat]]]
+        Hcore matrix in MO basis (1-el integrals)
     f_matrix: f_matrix: list[list[list[StrictFloat]]]
         Fock matrix/matrices
     j_matrix: list[list[list[StrictFloat]]]
         Coulomb integrals (2-el integrals)
     k_matrix: list[list[list[StrictFloat]]]
         Exchange integrals (2-el integrals)
+    twoelintegrals: TwoElectronIntegrals | None, default = None
+        Contains the available two electron integrals
     pointgroup: StrictStr
         Pointgroup of the molecule
     td_dft: TdDft | None default = None
@@ -57,9 +66,13 @@ class Molecule(GetItem):
     origin: tuple[StrictFloat, StrictFloat, StrictFloat]
     s_matrix: list[list[StrictFloat]] | None = Field(default=None, alias="s-matrix")
     h_matrix: list[list[StrictFloat]] | None = Field(default=None, alias="h-matrix")
+    t_matrix: list[list[StrictFloat]] | None = Field(default=None, alias="t-matrix")
+    v_matrix: list[list[StrictFloat]] | None = Field(default=None, alias="v-matrix")
+    hmo: list[list[list[StrictFloat]]] | None = None
     f_matrix: list[list[list[StrictFloat]]] | None = Field(default=None, alias="f-matrix")
     j_matrix: list[list[list[StrictFloat]]] | None = Field(default=None, alias="j-matrix")
     k_matrix: list[list[list[StrictFloat]]] | None = Field(default=None, alias="k-matrix")
+    twoelintegrals: TwoElectronIntegrals | None = Field(default=None, alias="2elintegrals")
     pointgroup: StrictStr | None = None
     td_dft: list[TdDft] | None = Field(None, alias="td-dft")
     densities: Densities | None = None