+++ src/opi/output/models/json/gbw/properties/two_electron_integral_element.py
@@ -1,0 +1,11 @@
+from pydantic import StrictFloat
+
+from opi.output.models.base.strict_types import StrictNonNegativeInt
+
+TwoElectronIntegralElement = tuple[
+    StrictNonNegativeInt,
+    StrictNonNegativeInt,
+    StrictNonNegativeInt,
+    StrictNonNegativeInt,
+    StrictFloat,
+]