+++ tests/examples/test_exmp038_integrals.py
@@ -31,6 +31,24 @@ def test_exmp038_integrals(example_input_file, tmp_path, json_files_exporter) ->
     assert hcore.dtype == np.float64
     assert hcore.ndim == 2
 
+    # kinetic
+    kinetic = output.get_int_kinetic(recreate_json=True)
+    assert isinstance(kinetic, np.ndarray)
+    assert hcore.dtype == np.float64
+    assert hcore.ndim == 2
+
+    # nuclear
+    nuclear = output.get_int_nuc_attr(recreate_json=True)
+    assert isinstance(nuclear, np.ndarray)
+    assert hcore.dtype == np.float64
+    assert hcore.ndim == 2
+
+    # hmo
+    hmo = output.get_int_hmo(recreate_json=True)
+    assert isinstance(hmo, np.ndarray)
+    assert hcore.dtype == np.float64
+    assert hcore.ndim == 2
+
     # f (although it rather is g)
     f = output.get_int_f(recreate_json=True)
     assert isinstance(f, np.ndarray)